@@ -13,10 +13,11 @@
         }
     ],
     "require": {
-        "php": ">=5.3.3"
+        "php": ">=5.5"
     },
     "require-dev": {
-        "phpunit/phpunit": "^4.7"
+        "phpunit/phpunit": "^4.7",
+        "squizlabs/php_codesniffer": "^2.3.1"
     },
     "autoload": {
         "psr-4": {
