@@ -25,7 +25,7 @@ class ViewWithHigherOrderSection
 
     public function bolder()
     {
-        return function($text, $renderer) {
+        return function ($text, $renderer) {
             return '<b>' . call_user_func($renderer, $text) . '</b>';
         };
     }
